moved the validate_schema method to ConfigFile class
